Move clockCrossing types into prci
