+++ src/main/scala/prci/ResetCrossingType.scala
@@ -4,8 +4,6 @@ package freechips.rocketchip.prci
 import org.chipsalliance.cde.config._
 import org.chipsalliance.diplomacy.lazymodule._
 
-import freechips.rocketchip.diplomacy.{CrossingType, HasDomainCrossing}
-
 trait HasResetDomainCrossing extends HasDomainCrossing { this: LazyModule =>
   type DomainCrossingType = ResetCrossingType
 }
