+++ src/main/scala/prci/package.scala
@@ -3,8 +3,7 @@
 package freechips.rocketchip
 
 import org.chipsalliance.diplomacy.nodes._
-
-import freechips.rocketchip.diplomacy.{ClockCrossingType, AsynchronousCrossing}
+import freechips.rocketchip.diplomacy.{BufferParams}
 
 package object prci
 {
@@ -18,4 +17,6 @@ package object prci
     case _: AsynchronousCrossing => async
     case _ => notasync
   }
+
+  implicit def noCrossing(value: NoCrossing.type): ClockCrossingType = SynchronousCrossing(BufferParams.none)
 }
