+++ src/main/scala/tilelink/CrossingHelper.scala
@@ -5,11 +5,10 @@ package freechips.rocketchip.tilelink
 import org.chipsalliance.cde.config._
 import org.chipsalliance.diplomacy.lazymodule._
 
-import freechips.rocketchip.diplomacy.{
+import freechips.rocketchip.prci.{
   AsynchronousCrossing, CrossingType, ClockCrossingType, NoCrossing,
-  RationalCrossing, CreditedCrossing, SynchronousCrossing
-}
-import freechips.rocketchip.prci.{ResetCrossingType, NoResetCrossing, StretchedResetCrossing}
+  RationalCrossing, CreditedCrossing, SynchronousCrossing,
+  ResetCrossingType, NoResetCrossing, StretchedResetCrossing}
 
 trait TLOutwardCrossingHelper {
   type HelperCrossingType <: CrossingType
